backend/pyproject.toml
[tool.poetry]
name = "newsletter"
version = "1.0.0"
description = "The backend processes that will run the analytics for Minerva app."
authors = ["Jack Bergin <jackcbergin@gmail.com>"]
license = "Copyright 2024, MINERVA, LLC. All rights reserved."

[tool.poetry.dependencies]
pydantic = "~1.10.8"
python = "^3.10"
azure-storage-file = "^2.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.23.3"
sqlalchemy = "^2.0.12"
pytest-cov = "^4.1.0"
numpy = "^1.24.3"
pandas = "^2.0.2"
tqdm = "^4.65.0"
openpyxl = "^3.1.2"
scipy = "^1.10.1"
datetime = "^5.1"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
websockets = "^10.0"
arxiv = "*"
youtube_transcript_api = "*"
openai = "*"
markdown = "*"
pytest = "^7.1.2"
praw = "^7.8.1"

[tool.poetry.dev-dependencies]
# basic code formatting
black = "^22.3.0"
isort = "^5.8.0"
pre-commit = "^2.13.0"
pre-commit-hooks = "^4.0.1"

# testing
pytest = "^7.1.2"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0"
pytest-memray = "^1.2.0"
pytest-sugar = "^0.9.5"
# linting
flake8 = "^3.9.2"

# type checking
mypy = "^1.5"

# docstring linting
darglint = "^1.8.0"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"

# test linting
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.5.0"

# versioning
bump2version = "^1.0.1"
ipykernel = "^6.13.1"
tox = "^3.24.4"

# profiling
memray = "^1.2.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.isort]
skip_glob = ["scripts/*"]
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]